requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
DateTime==5.5
Django==4.2.16
django-admin==2.0.2
django-axes==7.0.0
django-bootstrap5==24.3
django-environ==0.11.2
django-excel-response2==3.0.6
django-extensions==3.2.3
django-formtools==2.5.1
django-six==1.0.5
djangorestframework==3.15.2
docutils==0.21.2
et_xmlfile==2.0.0
excel-base==1.0.4
fonttools==4.54.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
isoweek==1.3.3
kiwisolver==1.4.7
matplotlib==3.9.2
mysqlclient==2.1.1
numpy==1.25.2
oauth2client==4.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyDMNrules==1.4.4
PyDrive==1.3.1
pyparsing==3.2.0
pySFeel==1.4.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rsa==4.9
screen==1.0.1
six==1.16.0
sly==0.5
sqlparse==0.5.1
statistics==1.0.3.5
tenacity==9.0.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
xlwt==1.3.0
zope.interface==7.1.1
